backend/pyproject.toml
[project]
name = "pipelines-gh"
version = "0.1.0"
description = "Multi-agent CI/CD pipeline migration to GitHub Actions"
requires-python = ">=3.11"
dependencies = [
    "github-copilot-sdk",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    "websockets>=14.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]
